refactor(net): rename child_abi module to materialize
